Move nav trigger logic to own component
